tagging FreeIPA 4.8.1

+++ ipaclient/install/ipa_client_samba.py
@@ -433,8 +433,9 @@ def uninstall(fstore, statestore, options):
     ipautil.remove_ccache(ccache_path=paths.KRB5CC_SAMBA)
 
     # Remove samba's configuration file
-    ipautil.remove_file(paths.SMB_CONF)
-    fstore.restore_file(paths.SMB_CONF)
+    if fstore.has_file(paths.SMB_CONF):
+        ipautil.remove_file(paths.SMB_CONF)
+        fstore.restore_file(paths.SMB_CONF)
 
     # Remove samba's persistent and temporary tdb files
     tdb_files = [
@@ -522,11 +523,25 @@ def run():
     if options.uninstall:
         if statestore.has_state("domain_member"):
             uninstall(fstore, statestore, options)
-            print(
-                "Samba configuration is reverted. "
-                "However, Samba databases were fully cleaned and "
-                "old configuration file will not be usable anymore."
-            )
+            try:
+                keys = (
+                    "configured", "hardening", "groupmap", "tdb",
+                    "service.principal", "smb.conf"
+                )
+                for key in keys:
+                    statestore.delete_state("domain_member", key)
+            except Exception as e:
+                print(
+                    "Error: Failed to remove the domain_member statestores: "
+                    "%s" % e
+                )
+                return 1
+            else:
+                print(
+                    "Samba configuration is reverted. "
+                    "However, Samba databases were fully cleaned and "
+                    "old configuration file will not be usable anymore."
+                )
         else:
             print("Samba domain member is not configured yet")
         return 0
@@ -624,7 +639,7 @@ def run():
             api.Command.service_del(api.env.smb_princ)
         except AttributeError:
             logger.error(
-                "Chosen IPA master %s does not have support to"
+                "Chosen IPA master %s does not have support to "
                 "set up Samba domain members", server,
             )
             return 1
